Backend/requirements.txt
asgiref==3.2.3
astroid==2.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
python-decouple==3.3
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
wrapt==1.11.2
